Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,9 +1,10 @@
 #include <stdio.h>
 int main(){
-    int radius;
-    scanf("%d",&radius);
-    int area = 3.14*radius*radius;
-    printf("Area: %d",area);
+    float radius;
+    scanf("%f",&radius);
+    float area = 3.14*(radius*radius);
+    printf("Area: %0.2f",area);
 
     return 0;
+
 }
